backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
scikit-learn
pandas
numpy
joblib
matplotlib
seaborn
xgboost
lightgbm
python-multipart